output: merge common functions
