+++ src/tr1/game/output.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "game/output/common.h"
 #include "game/output/draw.h"
 
 #include <libtrx/game/output.h>
