--- src/tr1/game/output/common.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#include "game/output/draw.h"
-#include "global/types.h"
-
-#include <libtrx/game/output/scene_compositor.h>
-#include <libtrx/game/output/shader.h>
-#include <libtrx/game/output/state.h>
-
-void Output_ApplyLevelSettings(void);
-void Output_ApplyRenderSettings(void);
